# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for LNCT GenZ Connect
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CategorySlug = typeof CategorySlug[keyof typeof CategorySlug];


export const CategorySlug = {
  find_friends: 'find_friends',
  hackathon_team: 'hackathon_team',
  study_group: 'study_group',
  roommate: 'roommate',
  project_collab: 'project_collab',
  other: 'other',
} as const;

export type PostStatus = typeof PostStatus[keyof typeof PostStatus];


export const PostStatus = {
  active: 'active',
  closed: 'closed',
  removed: 'removed',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];


export const RequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Category {
  slug: CategorySlug;
  label: string;
  description: string;
}

export type PostUrgency = typeof PostUrgency[keyof typeof PostUrgency];


export const PostUrgency = {
  casual: 'casual',
  urgent: 'urgent',
  looking_for_long_term: 'looking_for_long_term',
} as const;

export interface Post {
  id: number;
  category: CategorySlug;
  alias: string;
  title: string;
  description: string;
  skills: string[];
  status: PostStatus;
  viewCount: number;
  requestCount: number;
  createdAt: string;
  imageUrl?: string;
  imageUrls?: string[];
  urgency: PostUrgency;
  expiresAt?: string;
  contactNote?: string;
}

export type PostWithToken = Post & {
  ownerToken: string;
};

export type PostInputUrgency = typeof PostInputUrgency[keyof typeof PostInputUrgency];


export const PostInputUrgency = {
  casual: 'casual',
  urgent: 'urgent',
  looking_for_long_term: 'looking_for_long_term',
} as const;

export type PostInputExpiryDuration = typeof PostInputExpiryDuration[keyof typeof PostInputExpiryDuration];


export const PostInputExpiryDuration = {
  '24h': '24h',
  '3d': '3d',
  '7d': '7d',
  '30d': '30d',
  never: 'never',
} as const;

export interface PostInput {
  category: CategorySlug;
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  skills?: string[];
  imageUrl?: string;
  imageUrls?: string[];
  urgency?: PostInputUrgency;
  expiryDuration?: PostInputExpiryDuration;
  /** @maxLength 300 */
  contactNote?: string;
}

export type PostModerationUrgency = typeof PostModerationUrgency[keyof typeof PostModerationUrgency];


export const PostModerationUrgency = {
  casual: 'casual',
  urgent: 'urgent',
  looking_for_long_term: 'looking_for_long_term',
} as const;

export interface PostModeration {
  status: PostStatus;
  /**
     * @minLength 3
     * @maxLength 120
     */
  title?: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description?: string;
  category?: CategorySlug;
  urgency?: PostModerationUrgency;
  /** @maxLength 300 */
  contactNote?: string;
  skills?: string[];
}

export interface ConnectionRequestInput {
  /**
     * @minLength 5
     * @maxLength 1000
     */
  message: string;
  /**
     * Optional note on how the admin can reach the requester (e.g. preferred handle)
     * @maxLength 300
     */
  contactNote?: string;
}

export interface ConnectionRequestWithToken {
  id: number;
  postId: number;
  status: RequestStatus;
  message: string;
  createdAt: string;
  requesterToken: string;
}

export type AdminMessageToParty = typeof AdminMessageToParty[keyof typeof AdminMessageToParty];


export const AdminMessageToParty = {
  poster: 'poster',
  requester: 'requester',
} as const;

export interface AdminMessage {
  id: number;
  toParty: AdminMessageToParty;
  body: string;
  createdAt: string;
}

export interface MyPostIncomingRequest {
  id: number;
  status: RequestStatus;
  message: string;
  createdAt: string;
  messages: AdminMessage[];
}

export type MyPost = Post & {
  incomingRequests: MyPostIncomingRequest[];
};

export interface MyRequest {
  id: number;
  postId: number;
  postTitle: string;
  postCategory: CategorySlug;
  status: RequestStatus;
  message: string;
  createdAt: string;
  messages: AdminMessage[];
}

export interface AdminConnectionRequest {
  id: number;
  postId: number;
  postTitle: string;
  postCategory: CategorySlug;
  postAlias: string;
  status: RequestStatus;
  message: string;
  /** @nullable */
  contactNote: string | null;
  createdAt: string;
  messages: AdminMessage[];
}

export type AdminMessageInputToParty = typeof AdminMessageInputToParty[keyof typeof AdminMessageInputToParty];


export const AdminMessageInputToParty = {
  poster: 'poster',
  requester: 'requester',
} as const;

export interface AdminMessageInput {
  toParty: AdminMessageInputToParty;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  body: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface AdminStats {
  totalPosts: number;
  activePosts: number;
  pendingRequests: number;
  approvedRequests: number;
  postsToday: number;
  removedPosts: number;
}

export type ListPostsParams = {
category?: CategorySlug;
sort?: ListPostsSort;
search?: string;
};

export type ListPostsSort = typeof ListPostsSort[keyof typeof ListPostsSort];


export const ListPostsSort = {
  trending: 'trending',
  recent: 'recent',
} as const;

export type ListTrendingPostsParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type ListAdminPostsParams = {
status?: PostStatus;
};

export type ListAdminRequestsParams = {
status?: RequestStatus;
};