# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for LNCT GenZ Connect
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid password
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard summary counts for admin
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Not authenticated
  /admin/posts:
    get:
      operationId: listAdminPosts
      tags: [admin]
      summary: List all posts for moderation
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PostStatus"
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
  /admin/posts/{id}:
    patch:
      operationId: updateAdminPost
      tags: [admin]
      summary: Moderate a post (e.g. mark as spam/removed, restore, close)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostModeration"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
        "404":
          description: Not found
  /admin/requests:
    get:
      operationId: listAdminRequests
      tags: [admin]
      summary: List all connection requests for mediation
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RequestStatus"
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminConnectionRequest"
        "401":
          description: Not authenticated
  /admin/requests/{id}/approve:
    post:
      operationId: approveAdminRequest
      tags: [admin]
      summary: Approve a connection request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConnectionRequest"
        "401":
          description: Not authenticated
        "404":
          description: Not found
  /admin/requests/{id}/reject:
    post:
      operationId: rejectAdminRequest
      tags: [admin]
      summary: Reject a connection request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConnectionRequest"
        "401":
          description: Not authenticated
        "404":
          description: Not found
  /admin/requests/{id}/messages:
    post:
      operationId: sendAdminMessage
      tags: [admin]
      summary: Admin sends a mediated message to the poster or requester
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConnectionRequest"
        "401":
          description: Not authenticated
        "404":
          description: Not found
components:
  schemas:
...
    CategorySlug:
      type: string
      enum:
        [
          find_friends,
          hackathon_team,
          study_group,
          roommate,
          project_collab,
          other,
        ]
    PostStatus:
      type: string
      enum: [active, closed, removed]
    RequestStatus:
      type: string
      enum: [pending, approved, rejected]
...
    Post:
      type: object
      required:
        [
          id,
          category,
          alias,
          title,
          description,
          skills,
          status,
          viewCount,
          requestCount,
          createdAt,
          urgency,
        ]
      properties:
        id:
          type: integer
        category:
          $ref: "#/components/schemas/CategorySlug"
        alias:
          type: string
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/PostStatus"
        viewCount:
          type: integer
        requestCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        urgency:
          type: string
          enum: [casual, urgent, looking_for_long_term]
        expiresAt:
          type: string
          format: date-time
        contactNote:
          type: string
...
    PostModeration:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PostStatus"
        title:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          minLength: 10
          maxLength: 2000
        category:
          $ref: "#/components/schemas/CategorySlug"
        urgency:
          type: string
          enum: [casual, urgent, looking_for_long_term]
        contactNote:
          type: string
          maxLength: 300
        skills:
          type: array
          items:
            type: string
...
    AdminMessage:
      type: object
      required: [id, toParty, body, createdAt]
      properties:
        id:
          type: integer
        toParty:
          type: string
          enum: [poster, requester]
        body:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AdminConnectionRequest:
      type: object
      required:
        [
          id,
          postId,
          postTitle,
          postCategory,
          postAlias,
          status,
          message,
          contactNote,
          createdAt,
          messages,
        ]
      properties:
        id:
          type: integer
        postId:
          type: integer
        postTitle:
          type: string
        postCategory:
          $ref: "#/components/schemas/CategorySlug"
        postAlias:
          type: string
        status:
          $ref: "#/components/schemas/RequestStatus"
        message:
          type: string
        contactNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AdminMessage"
    AdminMessageInput:
      type: object
      required: [toParty, body]
      properties:
        toParty:
          type: string
          enum: [poster, requester]
        body:
          type: string
          minLength: 1
          maxLength: 1000
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
    AdminStats:
      type: object
      required:
        [
          totalPosts,
          activePosts,
          pendingRequests,
          approvedRequests,
          postsToday,
          removedPosts,
        ]
      properties:
        totalPosts:
          type: integer
        activePosts:
          type: integer
        pendingRequests:
          type: integer
        approvedRequests:
          type: integer
        postsToday:
          type: integer
        removedPosts:
          type: integer