# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for LNCT GenZ Connect
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySlug } from './categorySlug';
import type { PostStatus } from './postStatus';
import type { PostUrgency } from './postUrgency';

export interface Post {
  id: number;
  category: CategorySlug;
  alias: string;
  title: string;
  description: string;
  skills: string[];
  status: PostStatus;
  viewCount: number;
  requestCount: number;
  createdAt: Date;
  imageUrl?: string;
  imageUrls?: string[];
  urgency: PostUrgency;
  expiresAt?: Date;
  contactNote?: string;
}